testvisibility/src/main/java/com/crido/BasicCalculator.java
package com.crido;

public class BasicCalculator {
    int result;

    public BasicCalculator() {
        result = 0;
    }

    public int getResult() {
        return result;
    }

    public void setResult(int result) {
        this.result = result;
    }

    public void add(int a, int b) {
        result = a + b;
    }

    public void subtract(int a, int b) {
        result = a - b;
    }

    public void multiply(int a, int b) {
        result = a * b;
    }

    public void divide(int a, int b) {
        result = a / b;
    }

}
